fix gemm_cl

+++ src/blas.c
@@ -164,7 +164,7 @@ void smooth_l1_cpu(int n, float *pred, float *truth, float *delta, float *error)
     int i;
     for(i = 0; i < n; ++i){
         float diff = truth[i] - pred[i];
-        float abs_val = fabs(diff);
+        float abs_val = fabsf(diff);
         if(abs_val < 1) {
             error[i] = diff * diff;
             delta[i] = diff;
